refactor: move PropSetup class to new prop_setup.py
