+++ gmatpyplus/propagator/__init__.py
@@ -1,1 +1,2 @@
 from .propagator import *
+from .prop_setup import *
